fix imports post refactor transforms
